refactor(workflow): isolate viewer mode types module
