+++ packages/workflow/src/workflow-settings-viewer-mode.ts
@@ -1,0 +1,16 @@
+export type ViewerInputType = 'text' | 'textarea' | 'number' | 'boolean' | 'file';
+
+export interface ViewerInputField {
+	key: string;
+	label: string;
+	type: ViewerInputType;
+	required?: boolean;
+	placeholder?: string;
+	helpText?: string;
+	accept?: string;
+}
+
+export interface ViewerMode {
+	manual?: string;
+	inputs?: ViewerInputField[];
+}
